requirements.txt
fastapi
pydantic
requests
pymupdf
sentence-transformers
faiss-cpu
transformers
torch
uvicorn